+++ src/DualThemeProvider.tsx
@@ -1,7 +1,8 @@
-import { PaletteMode, Theme, createTheme } from '@mui/material';
+import { BrightnessMedium, DarkMode, LightMode } from '@mui/icons-material';
+import { Theme, createTheme } from '@mui/material';
 import { ReactNode } from 'react';
-import { DefaultPaletteModeProvider } from './DefaultPaletteModeContext';
 import { DualThemeProviderBase } from './DualThemeProviderBase';
+import { PaletteModeContextValue, PaletteModeProvider } from './PaletteModeContext';
 
 const defaultDarkTheme = createTheme({
   palette: {
@@ -15,16 +16,11 @@ const defaultLightTheme = createTheme({
   },
 });
 
-export interface DualThemeProviderProps {
+export interface DualThemeProviderProps extends Partial<PaletteModeContextValue> {
   /** Dark theme object */
   darkTheme?: Theme;
   /** Light theme object */
   lightTheme?: Theme;
-  /**
-   * Default palette mode
-   * @default 'auto'
-   */
-  defaultPaletteMode?: PaletteMode | 'auto';
   children: ReactNode;
 }
 
@@ -32,13 +28,23 @@ export function DualThemeProvider({
   darkTheme = defaultDarkTheme,
   lightTheme = defaultLightTheme,
   defaultPaletteMode = 'auto',
+  icons = {
+    auto: <BrightnessMedium />,
+    light: <LightMode />,
+    dark: <DarkMode />,
+  },
+  messages = {
+    auto: 'Auto',
+    light: 'Light',
+    dark: 'Dark',
+  },
   children,
 }: DualThemeProviderProps) {
   return (
-    <DefaultPaletteModeProvider value={defaultPaletteMode}>
+    <PaletteModeProvider value={{ defaultPaletteMode, icons, messages }}>
       <DualThemeProviderBase darkTheme={darkTheme} lightTheme={lightTheme}>
         {children}
       </DualThemeProviderBase>
-    </DefaultPaletteModeProvider>
+    </PaletteModeProvider>
   );
 }