+++ src/PaletteModeIconButton.tsx
@@ -1,38 +1,11 @@
-import { BrightnessMedium, DarkMode, LightMode } from '@mui/icons-material';
 import { IconButton, IconButtonProps } from '@mui/material';
 import { usePaletteMode } from './usePaletteMode';
 
 export function PaletteModeIconButton(props: IconButtonProps) {
-  const { userPaletteMode, setUserPaletteMode } = usePaletteMode();
-  let icon = <BrightnessMedium />;
-  switch (userPaletteMode) {
-    case 'light':
-      icon = <LightMode />;
-      break;
-    case 'dark':
-      icon = <DarkMode />;
-      break;
-    default:
-      icon = <BrightnessMedium />;
-  }
+  const { toggleUserPaletteMode, icon } = usePaletteMode();
 
   return (
-    <IconButton
-      color="inherit"
-      {...props}
-      onClick={() => {
-        switch (userPaletteMode) {
-          case 'light':
-            setUserPaletteMode('dark');
-            break;
-          case 'dark':
-            setUserPaletteMode('auto');
-            break;
-          default:
-            setUserPaletteMode('light');
-        }
-      }}
-    >
+    <IconButton color="inherit" {...props} onClick={toggleUserPaletteMode}>
       {icon}
     </IconButton>
   );