+++ src/index.tsx
@@ -1,3 +1,5 @@
 export * from './DualThemeProvider';
+export * from './PaletteModeButton';
 export * from './PaletteModeIconButton';
+export * from './PaletteModeListItemButton';
 export * from './usePaletteMode';