+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 # Changelog
 
+## 1.1.0 - 2023-07-01
+
+- Added `<PaletteModeButton/>` component
+- Added `<PaletteModeListItemButton/>` component
+- Added `messages` and `icons` props to `<DualThemeProvider/>` component
+- Added `icon`, `icons`, `message` and `messages` return to `usePaletteMode()` hook
+
 ## 1.0.0 - 2023-06-23
 
-- Added `<DualThemeProvider />` component
-- Added `<PaletteModeIconButton />` component
+- Added `<DualThemeProvider/>` component
+- Added `<PaletteModeIconButton/>` component
 - Added `usePaletteMode()` hook